Adding settings tab for SQL datasources.
