+++ packages/shared-core/src/helpers/integrations.ts
@@ -1,5 +1,18 @@
-import { SourceName } from "@budibase/types"
+import { Datasource, SourceName } from "@budibase/types"
 
 export function isGoogleSheets(type: SourceName) {
   return type === SourceName.GOOGLE_SHEETS
+}
+
+export function isSQL(datasource: Datasource): boolean {
+  if (!datasource || !datasource.source) {
+    return false
+  }
+  const SQL = [
+    SourceName.POSTGRES,
+    SourceName.SQL_SERVER,
+    SourceName.MYSQL,
+    SourceName.ORACLE,
+  ]
+  return SQL.indexOf(datasource.source) !== -1
 }
